chore(weakref): drop weak-napi library
It causes build errors on modern OSes due to
python issues. The library is not being updated
and Node now has builtin support since v14.9
