+++ package.json
@@ -3,11 +3,8 @@
   "version": "1.0.1",
   "description": "Throttle depending on function arguments.",
   "main": "index.js",
-  "browser": {
-    "./weakref": "./weakref-generic"
-  },
   "engines": {
-    "node": "10.x.x"
+    "node": ">=12"
   },
   "scripts": {
     "test": "npm -s run test:lint & npm -s run test:unit",
@@ -42,7 +39,6 @@
   },
   "homepage": "https://github.com/ambassify/throttle#readme",
   "dependencies": {
-    "hash-it": "^4.0.5",
-    "weak-napi": "^2.0.1"
+    "hash-it": "^4.0.5"
   }
 }
