+++ weakref.js
@@ -1,5 +1,17 @@
-const nodeMajorVersion = parseInt(process.versions.node.replace(/\..*/, ''));
+/* globals WeakRef */
 
-module.exports = nodeMajorVersion >= 12 ?
-    require('./weakref-generic') : // https://github.com/node-ffi-napi/weak-napi/issues/16
-    require('weak-napi');
+// Available since Node@14.9 or since Node@12 using --harmony-weak-refs flag
+const supportsWeakRef = (typeof WeakRef != 'undefined');
+
+if (supportsWeakRef) {
+    module.exports = function(v) { return new WeakRef(v); };
+    module.exports.get = function(v) { return v.deref(); };
+    module.exports.isDead = function(v) {
+        return (typeof v.deref() == 'undefined');
+    };
+} else {
+    module.exports = function(v) { return v; };
+    module.exports.get = function(v) { return v; };
+    module.exports.isDead = function() { return false; };
+    module.exports.isUnsupported = true;
+}
